% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette.distance.spearman}
\alias{silhouette.distance.spearman}
\title{Spearman distance from a matrix}
\usage{
silhouette.distance.spearman(X)
}
\arguments{
\item{X}{A matrix}
}
\description{
TODO
}
