% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette}
\alias{silhouette}
\title{silhouette}
\usage{
silhouette(X, cluster)
}
\arguments{
\item{X}{A numeric matrix.}

\item{cluster}{A data.frame
that contains clustering information with molecule and cluster}
}
\value{
\describe{
  \item{distance_df}{data.frame containing the cumputed distance between features.}
  \item{names}{list of chararcter containing the names of the features.}
  \item{data}{matrix containing original data.}
  \item{cluster_df}{data.frame containing the cluster information by features.}
  \item{coef.df}{data.frame containing ths silhouette coefficient by cluster.}
}
}
\description{
A more detailed description.
}
\examples{
data <- get_demo_silhouette()
res <- silhouette(X = data$data, cluster = data$cluster)

}
