% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette}
\alias{silhouette}
\title{silhouette}
\usage{
silhouette(X, cluster)
}
\arguments{
\item{X}{A numeric matrix of size NxP with feature in colnames.}

\item{cluster}{A numeric vector of sive ncol(X) containing the cluster
information by}
}
\description{
A more detailed description.
}
