% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette}
\alias{silhouette}
\title{silhouette}
\usage{
silhouette(X, cluster)
}
\arguments{
\item{X}{A numeric matrix.}

\item{cluster}{A data.frame
that contains clustering information,}
}
\value{
The sum of \code{x} and \code{y}.
\describe{
  \item{One}{First item}
  \item{Two}{Second item}
}
}
\description{
A more detailed description.
}
\examples{
data <- get_demo_silhouette()
res <- silhouette(X = data$data, cluster = data$cluster)
}
